Bug 769207 - part 5: Make `Element::GetAttributeChangeHit()` use `AttrModType` r=smaug,dom-core
Differential Revision: https://phabricator.services.mozilla.com/D263056

+++ layout/style/RestyleManager.cpp
@@ -3659,8 +3659,7 @@ void RestyleManager::AttributeChanged(Element* aElement, int32_t aNameSpaceID,
   auto changeHint = nsChangeHint(0);
   auto restyleHint = RestyleHint{0};
 
-  changeHint |= aElement->GetAttributeChangeHint(
-      aAttribute, static_cast<uint8_t>(aModType));
+  changeHint |= aElement->GetAttributeChangeHint(aAttribute, aModType);
 
   MaybeRestyleForNthOfAttribute(aElement, aNameSpaceID, aAttribute, aOldValue);
   MaybeRestyleForRelativeSelectorAttribute(aElement, aNameSpaceID, aAttribute,
